Add files via upload
This dramaturgical analysis examines Joe Orton's final play, What the Butler Saw, from the perspective of theatrical costume production. The focus of this analysis is on the narrative structure of the play and how repetition of action, undressing and cross-dressing, contributes to its effective development. That is, sexuality (in the play) as being unverifiable, in turns, reveals sexuality, rather than being located biologically, it remains itself a form of dress-up that can be altered freely.